pyproject.toml
[tool.poetry]
name = "hypertag"
version = "0.3.4"
description = "File organization made easy using tags"
readme = "README.md"
authors = ["Sean <sean-p-96@hotmail.de>"]
license = "MIT"
homepage = "https://github.com/SeanPedersen/HyperTag"
repository = "https://github.com/SeanPedersen/HyperTag"
keywords = ["tag", "tagging", "file-system", "file", "organization"]
classifiers = [
    "Environment :: Console"
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.6.1"
fire = "^0.3.1"
tqdm = "^4.55.0"
filetype = "^1.0.7"
watchdog = "^1.0.2"
python-igraph = "^0.8.3"
cairocffi = "^1.2.0"
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
PyPDF2 = "^1.26.0"
wordninja = "^2.0.0"
ftfy = "^5.8"
textract = "^1.6.3"
sentence-transformers = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
rope = "^0.18.0"
mypy = "^0.790"
bandit = "^1.7.0"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
